+++ drumigo/src/main/java/com/ftn/drumigo/domain/Vehicle.java
@@ -25,6 +25,9 @@ public class Vehicle {
     @JoinColumn(name = "driver_id", nullable = false, unique = true)
     private Driver driver;
     
+    @Column(name = "license_plate", nullable = false, unique = true, length = 20)
+    private String licensePlate;
+    
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "vehicle_type_id", nullable = false)
     private VehicleType vehicleType;
