+++ drumigo/src/main/java/com/ftn/drumigo/repository/RidePassengerRepository.java
@@ -3,7 +3,7 @@ package com.ftn.drumigo.repository;
 import com.ftn.drumigo.domain.users.Passenger;
 import com.ftn.drumigo.domain.Ride;
 import com.ftn.drumigo.domain.RidePassenger;
-import com.ftn.drumigo.domain.User;
+import com.ftn.drumigo.domain.users.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
