+++ frontend/src/app/features/profile/services/profile-api.service.ts
@@ -2,7 +2,7 @@ import { Injectable, inject } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable, map } from 'rxjs';
 import { ProfileData, VehicleInfo } from '../models/profile.model';
-import { environment } from '../../../../environments/environment.dev';
+import { environment } from '../../../../environments/environment';
 
 interface VehicleInfoResponse {
   id: number;
